requirements.txt
aioredis==1.3.1
amqp==2.6.1
anyjson==0.3.3
asgiref==3.2.10
asn1crypto==1.4.0
async-timeout==3.0.1
attrs==20.1.0
autobahn==20.7.1
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
billiard==3.6.3.0
celery==4.4.2
celery-once==3.0.1
certifi==2020.6.20
cffi==1.14.2
channels==2.4.0
channels-redis==3.0.1
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.0
daphne==2.5.0
Django==3.0.8
django-cors-headers==3.4.0
django-filter==2.3.0
django-filters==0.2.1
django-simple-history==2.11.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
fabric2==2.5.0
flower==0.9.5
gunicorn==20.0.4
hiredis==1.1.0
humanize==2.6.0
hyperlink==20.0.1
idna==2.10
importlib-metadata==1.7.0
incremental==17.5.0
invoke==1.4.1
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.11
Markdown==3.2.2
MarkupSafe==1.1.1
msgpack==1.0.0
msgpack-python==0.5.6
mysqlclient==2.0.1
nose==1.3.7
packaging==20.4
paramiko==2.7.1
paramiko-expect==0.2.8
Pillow==7.2.0
pip-review==1.1.0
prometheus-client==0.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-nmap==0.6.1
pytz==2020.1
PyYAML==5.3.1
redis==3.3.11
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
tornado==6.0.4
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
zipp==3.1.0
zope.interface==5.1.0
